Moved preprocessData-test test into react-devtools-shared folder
